Crate problem012

Problem 12 (Highly divisible triangular number)

Solution detail

We'll iterate up through the triangle numbers in order, and calculate how many divisors each one has, stopping when we find one with 500 divisors.

To calculate the number of divisors of the triangle number T(n) = n * (n+1) / 2 we'll need to have some precomputed prime numbers. How many? Well, the number of divisors of T(n) is:

d(n) * d((n + 1) / 2) if n is odd.

d(n / 2) * d(n + 1) if n is even.

This means that to factorise T(n), we need to know about the primes up to sqrt(n).

Finally, we need an upper bound on n to know in advance how many primes are required. The product of the first k primes has 2^k divisors, so there is definitely some n less than the product of the first 9 primes with at least 500 divisors. That means that we should sieve primes up to the square root of this product.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the first triangle number to have more than n divisors.